chore(refactor): refactor tests [#155772693]

+++ server/src/controllers/RecipesController.js
@@ -1,5 +1,5 @@
 import db from '../models/index';
-import { getPaginationMeta, check } from '../helpers/';
+import { getPaginationMeta } from '../helpers/';
 
 /**
  * RecipesController
@@ -113,7 +113,12 @@ export default class RecipesController {
 
     db.Recipes.findById(req.params.id)
       .then((recipe) => {
-        check(recipe, req, res, 'recipe');
+        if (!recipe) {
+          return res.status(404).jsend.fail({ message: 'The recipe does not exist' });
+        }
+        if (recipe.UserId !== req.user.userId) {
+          return res.status(401).jsend.fail({ message: 'You are not authorized to edit this recipe' });
+        }
         recipe.update({
           name: req.body.name || recipe.name,
           description: req.body.description || recipe.description,
@@ -135,7 +140,12 @@ export default class RecipesController {
   static deleteRecipe(req, res) {
     db.Recipes.findById(req.params.id)
       .then((recipe) => {
-        check(recipe, req, res, 'recipe');
+        if (!recipe) {
+          return res.status(404).jsend.fail({ message: 'The recipe does not exist' });
+        }
+        if (recipe.UserId !== req.user.userId) {
+          return res.status(401).jsend.fail({ message: 'You are not authorized to delete this recipe' });
+        }
         recipe.destroy({ force: true })
           .then(() => res.status(200).jsend.success({ message: 'Recipe has been successfully deleted' }))
           .catch(error => res.status(400).jsend.error(error));
@@ -196,7 +206,12 @@ export default class RecipesController {
   static editReview(req, res) {
     db.Reviews.findById(req.params.id)
       .then((review) => {
-        check(review, req, res, 'review');
+        if (!review) {
+          return res.status(404).jsend.fail({ message: 'The review does not exist' });
+        }
+        if (review.UserId !== req.user.userId) {
+          return res.status(401).jsend.fail({ message: 'You are not authorized to edit this review' });
+        }
         review.update({
           review: req.body.review || review.review
         })
@@ -215,7 +230,12 @@ export default class RecipesController {
   static deleteReview(req, res) {
     db.Reviews.findById(req.params.id)
       .then((review) => {
-        check(review, req, res, 'review');
+        if (!review) {
+          return res.status(404).jsend.fail({ message: 'The review does not exist' });
+        }
+        if (review.UserId !== req.user.userId) {
+          return res.status(401).jsend.fail({ message: 'You are not authorized to delete this review' });
+        }
         review.destroy()
           .then(() => res.status(200).jsend.success({ message: 'Review has been successfully deleted' }))
           .catch(error => res.status(400).jsend.error(error));
